users/requirements.txt
alembic==1.11.1
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==2.0.0
pydantic==1.10.2
pytest==7.2.0
httpx==0.24.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
